pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secaudit"
version = "0.1.0"
description = "Automated Security Audit Toolkit"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "responses>=0.24",
]

[project.scripts]
secaudit = "secaudit.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
